+++ kwimage/im_filter.py
@@ -13,6 +13,7 @@ def radial_fourier_mask(img_hwc, radius=11, axis=None, clip=None):
         https://docs.opencv.org/3.0-beta/doc/py_tutorials/py_imgproc/py_transforms/py_fourier_transform/py_fourier_transform.html
 
     Example:
+        >>> from kwimage.im_filter import *  # NOQA
         >>> import kwimage
         >>> img_hwc = kwimage.grab_test_image()
         >>> img_hwc = kwimage.ensure_float01(img_hwc)
@@ -106,6 +107,7 @@ def fourier_mask(img_hwc, mask, axis=None, clip=None):
         xdoctest -m kwimage.im_filter fourier_mask --show
 
     Example:
+        >>> from kwimage.im_filter import *  # NOQA
         >>> import kwimage
         >>> img_hwc = kwimage.grab_test_image(space='gray')
         >>> mask = np.random.rand(*img_hwc.shape[0:2])